% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_output.R
\name{add_percentage_signs}
\alias{add_percentage_signs}
\title{Add Percentage Signs to Data}
\usage{
add_percentage_signs(data, aggregation_method)
}
\arguments{
\item{data}{Data frame}

\item{aggregation_method}{Character string indicating the aggregation method}
}
\value{
Data frame with percentage signs added
}
\description{
Helper function to add "%" signs to percentage columns.
}
