% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{create_box_plot}
\alias{create_box_plot}
\title{Create Box Plot for Statistical Analysis}
\usage{
create_box_plot(
  data,
  ques,
  disag,
  analysis_type,
  title,
  color_primary,
  color_secondary,
  font_sizes = list(plot_title = 12, legend_title = 10, legend_text = 10, geom_text = 3,
    axis_title = 10)
)
}
\arguments{
\item{data}{A data frame containing the original survey data}

\item{ques}{Character string specifying the column name of the question to analyze}

\item{disag}{Character string specifying the disaggregation variable name}

\item{analysis_type}{Character string specifying the statistical method}

\item{title}{Character string for the chart title}

\item{color_primary}{Primary color}

\item{color_secondary}{Secondary color}
}
\value{
A ggplot2 object
}
\description{
Create Box Plot for Statistical Analysis
}
