% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{create_color_palette}
\alias{create_color_palette}
\title{Helper function to create color palette}
\usage{
create_color_palette(n, color_primary = "#730202", color_secondary = "#f27304")
}
\arguments{
\item{n}{Number of colors needed}

\item{color_primary}{Primary color}

\item{color_secondary}{Secondary color}
}
\value{
Vector of colors
}
\description{
Helper function to create color palette
}
