% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographics_table.R
\name{demographics_table}
\alias{demographics_table}
\title{Create Demographics table for surveys}
\usage{
demographics_table(vars_obj, survey_obj)
}
\arguments{
\item{vars_obj}{list of variables}

\item{survey_obj}{the survey object to use}
}
\value{
a tibble dataframe
}
\description{
Create Demographics table for surveys
}
\examples{
demographics_table(c("age","sex"), my_survey)
}
